fix: the API of es6-promisify is not the same as promisify-es6
Fixes a regression introduced by #896

`promisify-es6` was swapped for `es6-promisify` but the APIs of the
two modules are no the same so the NAT hole punching code broke.

Also changes port numbers to 0 in the tests as other processes can be
listening on those ports which causes spurious test failures.