feat: async peerstore backed by datastores
We have a peerstore that keeps all data for all observed peers in
memory with no eviction.

This is fine when you don't discover many peers but when using the
DHT you encounter a significant number of peers so our peer storage
grows and grows over time.

We have a persistent peer store, but it just periodically writes
peers into the datastore to be read at startup, still keeping them
in memory.

This change refactors the peerstore to use a datastore by default,
reading and writing peer info as it arrives.  It can be configured
with a MemoryDatastore if desired.

It was necessary to change the peerstore and *book interfaces to
be asynchronous since the datastore api is asynchronous.

This will need a follow up PR to libp2p-interfaces and others.