fix: abort dial memory leak
Protocol negotiation or encryption handshaking can hang indefinitely
when dialing peers with unreliable network connections.

When this happens the entry in the `_pendingDials` map in `src/dialer/index.js`
hangs around forever and the memory is never reclaimed.

Our timeout controller will fire the `abort` event on it's signal
so the connection will be torn down, but something has to be listening
for that event so pass the AbortSignal in to mss and noise.