fix: increase ping concurrency (#341)
After running an js-ipfs node for a week I noticed the lan routing table ping queue had 200k items in it (😱).  I think this is because we encounter lots of peers that are advertising their private addresses, they get added to the routing table, then we need to ping existing peers to see if we should evict them, which times out because the private addresses aren't routable.  We do this with `concurrency: 1` so it takes 10s to process each item in the queue but we are adding them to the queue at a much faster rate than that so it grows out of control.

Increasing the concurrency alleviates the queue pressure, so here we increase the default concurrency to 10 and expose it as a config option, as well as the timeout value to allow tuning by the user for their specific application.

We also add the ping queue size and current in-flight ping requests to the list of tracked metrics for the node.