fix: time out slow reads
There are a few places in the codebase where we send/receive data
from the network without timeouts/abort controllers which means
the user has to wait for the underlying socket to timeout which can
take a long time depending on the platform, if at all.

This PR ensures we can time out while running identify (both flavours),
ping and fetch.