fix!: do not auto-dial peers
Currently whenever a new peer is discovered we automatically dial
them if we are below the connection manager connection limit.

This is incredibly resource-intensive and does not result in high-quality
connections (e.g. KAD close, relays or supporting protocols we are
interested in).

Now that we can search the DHT for peers, this is no longer necessary
so remove the functionality.