fix: allow listening on duplicate addresses
Some listen addresses are turned into valid unique addresses when
listening on them, for example `/ip4/0.0.0.0/tcp/0`, so we should
be able to listen on multiple versions of addresses like these.

Currently we dedupe listen addresses before listening starts so
this isn't possible.

This PR removes the deduping and adds a test.