fix: when passed a multiaddr, only dial that multiaddr
When a user passes a multiaddr to `libp2p.dial`, only dial that
multiaddr instead of extracting the peer id from the multiaddr,
using it to look up every known address and dialing them all.

When passing a peer id to `libp2p.dial`, all addresses will be
dialed as before.

Fixes #1497