fix: circuitRelayTransport repeatedly connecting to relay
This fixes #1635 which causes the circuit relay to repeatedly connect
to a discovered relay. If the relay returns connection expiration in
seconds, we get a negative ttl when calculating
`expiration - new Date().getTime()`.
This caused the `addRelay` function to set the timeout at 0ms. This timeout
instantly triggers and calls `addRelay` recursively.