fix: if available, append target peer id to address book multiaddrs
Where we can load known multiaddrs for a peer during dial, append the
peer id of the remote to the multiaddrs.

This is for certain transports that require knowing the remote peer
like WebTransport, but those peers may have completed identify and
sent us addresses without a peer id appended.