fix: deriveKey in webkit linux (workaround) (#313)
The subtlecrypto implementation in WebKit on Linux doesn't like to derive a key from an empty imported key. This works on all other browsers, so it seems like the WebKit implementation is doing the wrong thing. Maybe worth opening a bug and writing a test for them. In the mean time here's a workaround.

This unblocks webkit testing in the interop tester (which runs on linux). This also lets folks use js-libp2p from webkit based linux browsers, although it doesn't seem like anyone else ran into this issue.

---------

Co-authored-by: Alex Potsides <alex@achingbrain.net>