fix: remove peer refresh
js-libp2p occasionally gets all peer routers to run a query for it's
own node id.

The only module that actually requires this is kad-dht in order to
insure it has peers when it runs a query and it does this anyway since
it's part of the spec.

https://github.com/libp2p/js-libp2p-kad-dht/pull/457 also updates kad-dht
to pause all executed queries until the self-query has run.

Consequently this is redundant and just creates unecessary work.