fix: wait for self-query to have run before running queries (#457)
In order to run a query we need DHT peers. The most common way to
find these is to run a self-query which happens after startup.

There's no way for the users to know when this has happened so they
end up putting arbitrary waits into the code so their queries have
a better chance of succeeding.

Instead just delay all queries until the self-query has run and
found some peers, then continue with the queries.