fix: build after updating version.ts
The default agent version is pulled from `version.ts` which is updated
before publish.  This is because dynamically reading from the `package.json`
file would cause it to get included in the browser bundle and requires
code contortions in ESM.

`version.ts` is updated in a `prepublishOnly` script to prevent PRs changing
this file accidentally, the fix here is to rebuild the ts files after
updating the version number, otherwise the published verison is still going
to be `0.0.0`.