fix: make address manager responsible for addresses
Refactors the transport manager to just start and stop transports.

Now the address manager is responsible for writing addresses into
the peer store.

Debounces writes to the peer store so we don't get lots of `self:peer:update`
events on startup as each transport starts to listen on configured
addresses.