fix: allow services to indicate their depdencies
Some services depend on others, for example circuit relay v2 and
kad-dht don't really work without identify so we need a way to
prompt the user to configure these dependencies.

The change is to allow extending the components a service requires
with a service map of the dependencies.

The components type in the `Libp2pInit` type is extended by the
`ServiceMap` generic type so when the user configures a node without
the required dependencies, TypeScript will fail to compile as the
components being passed to the factory function don't have type
overlap with the required components.