fix: use peer events in registrar
Refactors the registrar to use `peer:connect` and `peer:disconnect`
events instead of `connection:open` and `connection:close`.

The connection events are too low-level for the purpose and can
cause a race condition when multiple multiaddrs are dialed at the
same time.

When this happens the first connection survives and others are closed
but their corresponding `connection:open` event has already fired.

Interested parties registering topologies have their onConnect event
invoked with the soon-to-be-closed connection.

The `peer:connect` event happens later in the lifecycle by which
point the extra connection has been closed.