test: remove peer id fixtures
The peer id fixtures were introduced when all we had were RSA keys and
they were very slow to generate which slowed the tests down.

Now we have Ed25519 keys which are comparitivley fast to generate so
just use them.

This fixes some weirdness with timing around browser tests where the
relay server thinks it has reservations for peers that are reconnecting
due to the tests reusing peer ids.