feat: mark connections with limits as limited
Some connections have resources limits imposed on them, such as
circuit relay connections.

When this is the case, the connection will have a `.limited` boolean
property set to true.

By default any attempt to run a protocol over a limited connection
will throw (outgoing) or be reset (incoming), this is to prevent, for
example, bitswap exceeding the connection transfer limit and causing
the connection to be closed by the relay server when it should be
reserved for the WebRTC SDP exchange.

Protocols can opt-in to being run over limited connections by specifying
a flag during `libp2p.handle` (incoming) and `connection.openStream`
(outgoing).