feat: add reconnect threshold for peers that we have failed to dial
When auto-dialing peers, if we have failed to dial them recently there's
little point in redialing them.

Adds a `autoDialPeerRetryThreshold` which is a value in ms.

If we have attempted to dial a peer and that dial attempt failed but
we are under our min connection count, do not auto dial the peer within
the retry threshold.

Defaults to 10 minutes.