fix: only dial one address at a time for peers
The observation for dialing is that for a given peer with multiple
addresses, for a given class of address (e.g. public or private),
either the first dial succeeds or they all fail, so there's not
much point dialing all addresses together since in either case you
waste resources either cancelling redundant dials or dialing addresses
that fail.

We can't do much about dialing addresses that fail but we can
reduce the number of redundant dials so with this change we only ever
dial one multiaddr at a time.