fix(@libp2p/webrtc): delay closing datachannels
Tactical fix until we can update the spec to coordinate closing
datachannels.

Calling `.close` on a datachannel causes a stream reset - depending
on the implementation there's a good chance any data waiting to
be transmitted will be lost.

As a workaround when closing a datachannel wait until it's
`.bufferedAmount` property is `0`, then wait a bit longer before
closing the channel.

This is because although Firefox will have sent all the data by
this point, Chrome may have not.

This should be removed when the spec is updated.