fix(libp2p): opt-in to toplogy notifications on transient connections
Adds a `notifyOnTransient` option when registering a network topology
to opt-in to being notified when peers that support the registered
protocol connect over transient connections.  False by default.

The logic has been switched to notify after identify has completed
rather than after the first connection for a peer has been opened.

The side effect here is that if `notifyOnTransient` is true, and the
peer ends up opening a direct connection (for example they dial us
via circuit relay, open a stream to do the WebRTC SDP exchange, then
open a WebRTC connection), identify will run on the second connection
so the topology will receive two notificiations.

This is not a breaking change since the previous behaviour would have
been to only notify on the transient connection, which is probably a
bug as you can't do data-heavy things like bitswap over transient
connections so is undesirable.

Fixes #2036