fix: add FIN_ACK message to WebRTC streams to prevent early closing
Chrome does not always send all messages after closing a datachannel
even if `bufferedAmount` is `0` before closing.

This PR adds a `FIN_ACK` message that is sent in reply to a `FIN`
message - because all messages are send in-order, when this is
received we know the remote has recieved all of our data messages
and it's safe to close the channel.