fix!: remove min/max from topologies (#2158)
Once upon a time these options were intended to cause libp2p to
close connections or search for more peers that support a given
protocol but it was never implemented.

Remove the options since they don't do anything, they may be
restored in future if the functionality is ever required.