fix!: remove protocols from PeerInfo
A [PeerInfo](https://docs.libp2p.io/concepts/fundamentals/peers/#peer-info)
is a libp2p object that combines a PeerID and some Multiaddrs.

We also add a list of protocols. This was a mistake because protocols are
exchanged during Identify but the PeerInfo object is used for peer discovery.

This is evident because we set the protocol list to an empty array everywhere.

PeerInfo is useful for exchanging peer data with other nodes, if we need a
more fleshed-out peer representation we'd use the `Peer` interface from
the peer store.

BREAKING CHANGE: the `.protocols` property has been removed from the `PeerInfo` interface