fix: allow DHT self-query to time out (#2169)
When we run our first self-query we first check the routing table to
see if it has peers.

If not we wait for some peers to be discovered before continuing.

The change here is to use the general query AbortSignal to abort
waiting for peers to be discovered.  If they are not discovered allow
other queries to continue and fail with eaiser to understand error
messages.

Also fixes some logging typos and makes error messages more understandable.