fix: rename event emitter class (#2173)
We add types to a `EventEmitter` class which extends the plain js `EventTarget`.

The name is chosen to give familiarity to node developers but it's not generally a good idea.

This refactor:

1. Exports a new `TypedEventTarget` interface which adds types to `EventTarget`
2. Renames `EventEmitter` to `TypedEventEmitter` to draw a distinction between them
3. Re-exports `TypedEventEmitter` as `EventEmitter` to preserve backwards compatibility

In the future all consuming code should rely on the `TypedEventTarget` interface while implementing code uses `TypedEventEmitter` as an implementation of a `TypedEventEmitter`.

By depending on the interface and not the implementation consuming code is isolated from problems that arise when two versions of `@libp2p/interface` is in the dependency tree and the event subsystems would otherwise be compatible due to type overlap.

This manifests as an error message about incompatible implementations of the private `#listeners` field in EventEmitter.

Backwards compatibility is achieved by exporting the `TypedEventEmitter` class as the older `EventEmitter` name.  This should be removed in a future PR to the v1.0 branch once external modules (e.g. Gossipsub) have been updated to use the new `TypedEventEmitter` symbol.

## Change checklist

- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation if necessary (this includes comments as well)
- [ ] I have added tests that prove my fix is effective or that my feature works