fix: allow dialing a peer when we only have transient connections
If we dial a peer that we already have a connection to, we return that
connection instead of dialing.

This creates a problem when we are trying to make a direct connection
to a peer that we already have a relay connection to since the relay
connection is returned.

The fix here is to allow additional dials to peers when we only have
relay connections.

If a direct connection exists, we will still return it instead of dialing.