fix: make initiator webrtc address dialable
Updates the reported remote address of webrtc connections to be the
address that was dialed, so it can be treated as an opaque value and
shared with other interested parties.

The recipient remote address is just `/webrtc/p2p/QmInitiator` as there
is no guarantee that the intiator has a reservation anywhere.