fix: close early WebRTC streams properly
A race condition exists where a remote can start opening streams
over WebRTC before we have finished the connection upgrade.

When this happens we need to be able to gracefully close the stream
the same as ones opened after the upgrade.

This PR makes the action taken on stream close reconfigurable after
the muxer factory has been created.