refactor: remove abortable iterator
We use an `abortable-iterator` to race an abort signal against an
incoming for streams and connections so we do not wait on slow connections
forever.

Unfortunately this introduces so much latency it becomes a performance
bottleneck, and in all cases we can just add an `abort` listener to the
incoming abort signal and close the underlying stream if it fires.

Removing the latency added by the abortable iterator reveals that tests using
it-pair, it-pushable and friends have many timing issues so it's necessary
to ensure we are supplying data to streams as it is required.

This opens up the possibility of having backpressure applied to protocol
streams by the underlying transports which isn't currently possible.