feat: allow joining jobs in peer queues
When we have a PeerJobQueue, the jobs are usually expensive but the
result is transferable - opening a connection, etc.

This PR adds a `joinJob` method to `PeerJobQueue` - if a job is queued
for the passed `PeerId`, a promise is returned that resolves/rejects
based on the outcome of the job.