feat!: use single DHT only by default
We have a weird dual-DHT setup whereby we have a DHT for public
addresses (amino, `/ipfs/kad/1.0.0`) and a DHT for private addresess
(`/ipfs/lan/kad/1.0.0`).

This is an artefact of the previous libp2p setup whereby there could
only be a single DHT implementation at once.

Now we have services we can configure an amino DHT service and if
desired an additional lan-only DHT, or the user can just use amino
and not pay the overhead of running an extra DHT.

Most content is resolved via the public DHT so running the lan-only
DHT gives the user no benefit.

BREAKING CHANGE: the `kadDHT` function returns a single DHT - see the readme for how to configure amino/lan as before