fix: return self in FIND_NODE for self
Partial revert of #2499

If a node is queried for it's own peer id, return it's own peer
info.

This is necessary because since https://github.com/libp2p/go-libp2p-kad-dht/pull/820
go-libp2p-kad-dht won't add a peer to it's routing tables that
doesn't have any DHT peers that are KAD-futher from it's own ID
already.