feat: add cuckoo filter (#2510)
Add a filter type that lets us efficiently remove items as well as
add them.

It would be better to use the `bloom-filters` module at this point
but it adds 50KB+ to browser bundles for very simple use cases so
it's not suitable.  We can revisit if https://github.com/Callidon/bloom-filters/issues/70
is ever resolved.