fix: close connections after use
Historically we've not tried to close connections after they've been
opened because we let the connection pruner do that when it's
required.

This leads to high resource usage and requires the pruner to be able
to select the best connections for pruning which without the context
of their intended use is a guessing game.

- Where protocols open connections that are not expected to be long-lived, try to close them if we are done
- Check the number of open streams on a connection before closing, if there are none or they are only speculative protocols, close the connection
- Sort connections by number of open streams in the pruner before pruning