feat: check service dependencies on startup
Allows services to optionally define the capabilities they provide to the
rest of libp2p and also the capabilities they require from other services.

This allows, for example, the `WebRTC` transport to require the `CircuitRelay`
transport to be present, or `KAD-DHT` (or anything that uses a topology)
to require the identify protocol.

Fixes https://github.com/libp2p/js-libp2p/issues/2263
Refs https://github.com/libp2p/js-libp2p/discussions/2135