feat!: remove autodialer
The autodialer is a feature from an older time when the DHT was less
reliable and we didn't have things like the random walk component.

There's not a lot of benefit in opening connections to any old peer,
instead protocols now have better ways of targetting the kind of
peers they require.

Actively dialing peers harms platforms where connections are extremely
expensive such as react-native so this feature has been removed.

Closes #2621

BREAKING CHANGE: the autodialer has been removed as well as the corresponding config keys