feat: add connection monitor
Adds a connection monitor that periodically ensures remote peers are
still online and contactable by trying to send a single byte via the
ping protocol, and sets the `.rtt` property of the connection to how
long it took.

If the ping protocol is not supported by the remote, it tries to infer
the round trip time by how long it took to fail.

If the remote is unresponsive or opening the stream fails for any
other reason, the connection is aborted with the throw error.

It's possible to configure the ping interval, how long we wait before
considering a peer to be inactive and whether or not to close the
connection on failure.

Closes #2643