fix: vendor in debug module
debug is a CJS module with CJS dependencies.

It will be an ESM module in it's [next major release](https://github.com/debug-js/debug/issues/656),
however that release issue has been open for over five years so it
may never arrive.

This PR vendors in the debug module as typescript so we can move
on from CJS.