fix: accept custom ping protocol prefix in connection monitor (#2667)
The new connection monitor currently uses the hard-coded default ping protocol `'/ipfs/ping/1.0.0'`. The ping service lets the developer override the `ipfs` component with a custom protocol prefix, and some apps are already doing this. This PR adds support for a `protocolPrefix` option in `ConnectionMonitorInit` that matches the option in `PingServiceInit`.

---------

Co-authored-by: Alex Potsides <alex@achingbrain.net>