fix: prevent duplicate connections to multiaddr
After opening a new connection, check if there is an existing connection
to the same peer. If the existing connection is not limited or the new
connection is also limit, return the existing connection.

This will prevent multiple connections to the same peer in cases where
the multiaddr being dialed does not contain a peer id.