fix: track closest peers separately from main routing table
The routing table is a balance trie where the path to the leaf node
storing the contact is derived from the prefix of the kad id of the
contact.

This makes it great for starting a query because we can quickly find
contacts in the kad-vicinity of the target, but it's less good for
knowing peers that are in our kad-vicinity, since the bits that make
us kad-close to another peer might not be in the prefix.

Instead, use a peer distance list that we update whenever a peer
successfully completes a `PING` operation.  Periodically check this
list and tag the closes peers with `KEEP_ALIVE` so we maintain
connections to them, which will ensure we propagate changes in our
PeerInfo to those peers most likely to answer `FIND_PEER` queries
for our data.