fix: remove peer cache
In cases where the peer store has grown very large (e.g. when a
well connected node has been online for a long time), the peer store
can get so large that this cache causes abnormal memory use.

The cache was added without any profiling, remove the cache for now,
we can add it back if it actually solves a problem