fix: only close stream if it is open
Adds a guard to the `.close` operation similar to the one on
`.abort` that ensures we only close the stream if it is open.

The individual `.closeRead`/`.closeWrite` operations already guard
on the read/write status of the stream so there's no functional
change, we just avoid a bit more async work as those methods
return promises.