fix: gracefully handle remote stream closure during ping (#2822)
Fixes #2770. This issue happens because the ping protocol does not handle the case where the remote peer closes the stream and with the introduction of `connectionMonitor` this became an issue. It opens a stream for the ping protocol, sends a ping, gets the response then closes the stream. The local peer throws an exception and triggers a yamux reset.

---------

Co-authored-by: privacyguard <privacyguard@users.noreply.github.com>